LayoutTests/fast/selectors/selector-filter-hashes-expected.txt

Test how many selector filter hashes in each category we extract from a selector.
selector (ids, classes, tags, attributes)
#a (0, 0, 0, 0)
.a (0, 0, 0, 0)
a (0, 0, 0, 0)
[a] (0, 0, 0, 0)
[a=foo] (0, 0, 0, 0)
#a target (1, 0, 0, 0)
.a target (0, 1, 0, 0)
a target (0, 0, 1, 0)
[a] target (0, 0, 0, 1)
[a=foo] target (0, 0, 0, 1)
#a > target (1, 0, 0, 0)
.a > target (0, 1, 0, 0)
a > target (0, 0, 1, 0)
[a] > target (0, 0, 0, 1)
#a + target (0, 0, 0, 0)
.a + target (0, 0, 0, 0)
a + target (0, 0, 0, 0)
[a] + target (0, 0, 0, 0)
#a ~ target (0, 0, 0, 0)
.a ~ target (0, 0, 0, 0)
a ~ target (0, 0, 0, 0)
[a] ~ target (0, 0, 0, 0)
b#a target (1, 0, 1, 0)
b.a target (0, 1, 1, 0)
b[a] target (0, 0, 1, 1)
c[d].e b#a target (1, 1, 2, 1)
c[d].e b.a target (0, 2, 2, 1)
c[d].e b[a] target (0, 1, 2, 2)
c[d].e b#a + target (0, 1, 1, 1)
c[d].e b.a + target (0, 1, 1, 1)
c[d].e b[a] + target (0, 1, 1, 1)
c[d].e + b#a target (1, 0, 1, 0)
c[d].e + b.a target (0, 1, 1, 0)
c[d].e + b[a] target (0, 0, 1, 1)
c[d].e b#a ~ target (0, 1, 1, 1)
c[d].e b.a ~ target (0, 1, 1, 1)
c[d].e b[a] ~ target (0, 1, 1, 1)
:not(a) (0, 0, 0, 0)
:not(a) target (0, 0, 0, 0)
:not(a, .b) target (0, 0, 0, 0)
:has(a) (0, 0, 0, 0)
:has(a) target (0, 0, 0, 0)
:has(a, .b) target (0, 0, 0, 0)
:is(a) (0, 0, 0, 0)
:is(a) target (0, 0, 1, 0)
:is(a, .b) target (0, 0, 0, 0)
:is(a .b) target (0, 1, 1, 0)
:is(a + .b) target (0, 1, 0, 0)
:is(a .b) + target (0, 0, 0, 0)
#c :is(a, .b) target (1, 0, 0, 0)
#c :is(a .b) target (1, 1, 1, 0)
